refactor: move manage room components in their own folder
