+++ pages/explore/manage-room/TemplateSelect.js
